server jwt

+++ server/src/resolvers.ts
@@ -1,7 +1,11 @@
 import { PrismaClient } from '@prisma/client';
+import { AuthenticationError } from 'apollo-server';
 import { head, isNil, isEmpty } from 'ramda';
+import bcrypt from 'bcrypt';
+import jwt from 'jsonwebtoken';
 
 interface ResolverContext {
+  userId: string | null;
   prisma: PrismaClient;
 }
 
@@ -20,12 +24,16 @@ const resolvers = {
 
   Mutation: {
     createWordDoc: async (_: any, args: any, context: ResolverContext) => {
+      if (isNil(context.userId)) throw new AuthenticationError('not authentication');
+
       const word = await context.prisma.wordDoc.create({
         data: { ...args.word, prefix: head(args.word.word).toUpperCase() },
       });
       return word;
     },
     updateWordDoc: async (_: any, args: any, context: ResolverContext) => {
+      if (isNil(context.userId)) throw new AuthenticationError('not authentication');
+
       const word = await context.prisma.wordDoc.update({
         where: {
           id: args.id,
@@ -35,6 +43,8 @@ const resolvers = {
       return word;
     },
     deleteWordDoc: async (_: any, args: any, context: ResolverContext) => {
+      if (isNil(context.userId)) throw new AuthenticationError('not authentication');
+
       const word = await context.prisma.wordDoc.delete({
         where: {
           id: args.id,
@@ -43,6 +53,8 @@ const resolvers = {
       return word;
     },
     increaseWordReviewCount: async (_: any, args: any, context: ResolverContext) => {
+      if (isNil(context.userId)) throw new AuthenticationError('not authentication');
+
       const word = await context.prisma.wordDoc.update({
         where: {
           id: args.id,
@@ -54,6 +66,44 @@ const resolvers = {
         },
       });
       return word;
+    },
+
+    signup: async (_: any, args: any, context: ResolverContext) => {
+      const password = await bcrypt.hash(args.password, 10);
+
+      const user = await context.prisma.user.create({
+        data: {
+          ...args,
+          password,
+        },
+      });
+
+      const token = jwt.sign({ userId: user.id }, process.env.JWT_SECRET_KEY as string);
+
+      return {
+        token,
+        user,
+      };
+    },
+
+    login: async (_: any, args: any, context: ResolverContext) => {
+      const user = await context.prisma.user.findUnique({ where: { email: args.email } });
+
+      if (!user) {
+        throw new Error('No such user found');
+      }
+
+      const valid = await bcrypt.compare(args.password, user.password);
+      if (!valid) {
+        throw new Error('Invalid password');
+      }
+
+      const token = jwt.sign({ userId: user.id }, process.env.JWT_SECRET_KEY as string);
+
+      return {
+        token,
+        user,
+      };
     },
   },
 };
